feat: add hot wrapper utilities
